json/xml delete limited

+++ bookstore/books/views.py
@@ -2,12 +2,17 @@ from django.shortcuts import render
 from django.contrib.auth import login
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from django.http import JsonResponse
 from rest_framework import authentication, permissions
 from django.contrib.auth.models import User
-from .models import CustomUser
+from .models import CustomUser,Book
 from .serializers import CustomUserSerializer,BookSerializer
 from .serializers import LoginSerializer
 from rest_framework.authtoken.models import Token
+from rest_framework_xml.renderers import XMLRenderer
+from rest_framework_xml.parsers import XMLParser
+from django.http import HttpResponse
+
 class AllUsers(APIView):
     """
     View to list all users in the system.
@@ -17,6 +22,8 @@ class AllUsers(APIView):
     """
     #authentication_classes = [authentication.TokenAuthentication]
     #permission_classes = [permissions.IsAdminUser]
+    renderer_classes = [XMLRenderer, ]
+    #parser_classes = (XMLParser,)
 
     def get(self, request, format=None):
         """
@@ -24,7 +31,11 @@ class AllUsers(APIView):
         """
         users = CustomUser.objects.all()
         serializer = CustomUserSerializer(users, many=True)
-        return Response(serializer.data)
+        if request.content_type == 'application/json':
+            return JsonResponse(serializer.data,safe=False)
+        if request.content_type == 'application/xml':
+            return Response(serializer.data)
+           
     
     def post(self, request, format=None):
         data=request.data.copy()
@@ -59,17 +70,25 @@ class UserDetailsView(APIView):
     
     def delete(self, request, pk, format=None):
         user = CustomUser.objects.get(pk=pk)
-        user.delete()
-        return Response(status=204)
+        if user:
+            if user==self.request.user:
+                user.delete()
+                return Response(status=204)
+            return Response({"message":"Not allowed to delete"})
+        return Response(status=400)
+        
     
 class AllBooks(APIView):
      def get(self, request, format=None):
         """
         Return a list of all users.
         """
-        AllBooks = AllBooks.objects.all()
+        AllBooks = Book.objects.all()
         serializer = BookSerializer(AllBooks, many=True)
-        return Response(serializer.data)
+        if request.content_type == 'application/json':
+            return JsonResponse(serializer.data,safe=False)
+        if request.content_type == 'application/xml':
+            return Response(serializer.data)
     
      def post(self, request, format=None):
         data=request.data
@@ -78,6 +97,38 @@ class AllBooks(APIView):
             serializer.save()
             return Response(serializer.data, status=201)
         return Response(serializer.errors, status=400)
+
+class BookDetailsView(APIView):
+    def get(self, request, pk, format=None):
+        book = Book.objects.get(pk=pk)
+        serializer = BookSerializer(book)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        book = Book.objects.get(pk=pk)
+        serializer = BookSerializer(book, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
+
+    def patch(self, request, pk, format=None):
+        book = Book.objects.get(pk=pk)
+        serializer = Book(book, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
+    
+    def delete(self, request, pk, format=None):
+        book = Book.objects.get(pk=pk)
+        if book:
+            if book.author==self.request.user:
+                book.delete()
+                return Response(status=204)
+            return Response({"message":"Not allowed to delete, not author"})
+        return Response(status=400)
+    
      
 
 class LoginView(APIView):
